Return value from callable to work around VM issue
Due to https://github.com/dart-lang/sdk/issues/56064, Dart stable 3.4
reports an exception on native calls involving `NativeCallable`s
returning `null` (which they do for `void`).
This exception is not critical as it only occurs when a debugger is
attached, but it's still annoying and easy to work around by returning
a bogus value then dropped.

Closes #242