Wasm: Use Dart externrefs (#328)
Making Dart functions callable from C code compiled to WebAssembly used to work by:

1. Registering the Dart function in a global `Map<int, Function>`.
2. Passing the key to C as a `void*` context pointer.
3. Adding a static entrypoint callback receiving that context, looking up the function, and finally calling it.

Those are a lot of steps! Ideally, we'd be able to pass `JSFunction`s as `externref`s and then do `extern.convert_any` followed by a cast to functions and `call_ref`. I couldn't convince clang to generate that though.

So instead, we keep the global entryoints. But instead of maintaining a map in Dart, we could:

1. Convert a Dart object to an `ExternalDartReference`.
2. Pass that reference as an `externref` into a C function.
3. As long as C code needs access to that reference, it stores it in a wasm `table`.
4. Static entrypoint callbacks now receive the `ExternalDartReference` as their first argument, so the lifetime is maintained in C alone.